conditionally register endpoints, skip tests for disabled endpoints
