+++ backend.example.env
@@ -4,6 +4,12 @@ API_DOMAIN=http://localhost:8000
 CELERY_BROKER_URL=redis://redis:6379/0
 CELERY_RESULT_BACKEND=redis://redis:6379/0
 
+# Feature flags for optional modules
+ENABLE_BREEDBASE=false
+ENABLE_CAMPAIGNS=false
+ENABLE_IFORESTER=false
+ENABLE_STAC=false
+
 # separate each extension by comma
 EXTENSIONS=
 # container mount point for external storage (see .env for host mount point)
