+++ backend/app/tests/api/api_v1/test_contact.py
@@ -4,8 +4,10 @@ from sqlalchemy.orm import Session
 
 from app.core.config import settings
 from app.core.mail import fm
+from app.tests.conftest import pytest_requires_mail
 
 
+@pytest_requires_mail
 def test_email_contact_message_to_support(
     client: TestClient, db: Session, normal_user_access_token: str
 ) -> None:
@@ -37,6 +39,7 @@ def test_email_contact_message_to_support(
         )
 
 
+@pytest_requires_mail
 def test_email_contact_message_with_too_few_characters(
     client: TestClient, db: Session, normal_user_access_token: str
 ) -> None:
@@ -48,6 +51,7 @@ def test_email_contact_message_with_too_few_characters(
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
 
+@pytest_requires_mail
 def test_email_contact_message_with_too_many_characters(
     client: TestClient, db: Session, normal_user_access_token: str
 ) -> None:
@@ -67,6 +71,7 @@ def test_email_contact_message_with_too_many_characters(
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
 
+@pytest_requires_mail
 def test_email_contact_message_without_email_enabled_on_instance(
     client: TestClient, db: Session, normal_user_access_token: str
 ) -> None:
